🔧 chore: extract SwiftUI-facing extensions to DesignTokens+SwiftUI.swift
Splitting keeps DesignTokens.swift under the 400-line file_length cap
after the textStyle modifier and Dynamic Type breadcrumb. Also
pre-aligns with the eventual SPM split — tokens (data) stay in one
module, SwiftUI bindings (Color aliases + View modifier) in another.

No behavior change — pure file relocation.
